Update buildspec.yml
Prepending the call to ./create-new-task-def.sh with a specific call to /bin/sh in order to overcome failures when pushing to AWS CodeCommit from a Windows box running Git for Windows (and maybe others).  When pushed from Git for Windows, create-new-task-def.sh appears to loses the execute bit and this call to it fails with "permission denied" and "exit status 126" even though Git for Windows shows 755 permissions.